requirements.txt
torch
sentencepiece
transformers
fastapi
starlette
aiofiles
python-multipart
Pillow
mangum